destructuring index.js into components, moving all <p> text to markdown libraries
