+++ package.json
@@ -9,6 +9,8 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@mdx-js/loader": "^2.1.1",
+    "@next/mdx": "^12.1.6",
     "@popperjs/core": "^2.11.5",
     "bootstrap": "5.1.3",
     "next": "12.1.6",
