--- pages/Rentals/Rentals.js
@@ -1,9 +1,0 @@
-import react from 'react'
-
-import React from 'react'
-
-const Rentals = () => {
-  return <div>rentals</div>
-}
-
-export default Rentals
