+++ pages/home/rental.mdx
@@ -1,0 +1,4 @@
+<p>
+  Our rental program provides you a flexible and sensible path towards ownership
+  while saving you money on regular maintenance and repairs.
+</p>
