+++ pages/rentals/Info.js
@@ -1,0 +1,13 @@
+import React from 'react'
+
+import Info from './Info.mdx'
+
+const Rentals = () => {
+  return (
+    <div className="text-dark container bg-secondary">
+      <Info />
+    </div>
+  )
+}
+
+export default Rentals
