+++ pages/rentals/Info.mdx
@@ -1,0 +1,30 @@
+# How It Works
+
+&nbsp;
+
+### Monthly subscription
+
+- No minimum contract.
+- No cap on how long you can build equity with our rental program.
+  &nbsp;
+
+### A <small>small</small> portion of the payment every month covers maintenance and repairs
+
+- Significantly less than what you can expect to pay to maintain an actively used instrument.
+- Covers replacing spoiled ('false') or broken strings, providing a fresh bow when it is time to rehair, minor cosmetic wear, and expected repairs.
+  &nbsp;
+
+### The remainder is held as equity towards a purchase
+
+- 100% of your payment after maintenance and tax
+- Again, there is no cap on how much equity you can build
+- For each instrument, there is a flat mininum purchase price set at an entry level threshold for an instrument worth maintaining.
+- You are able to convert your active rental into equity at any time toward any outfit
+
+&nbsp;
+
+# Why Renting Is In Your Best Interest
+
+&nbsp;
+
+# Our Program Grows With You
